refactor: centralize nextauth configuration
