+++ motostix/src/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,1 @@
-// src/app/api/auth/[...nextauth]/route.ts
-import { handlers } from "@/auth"; // Import from auth.ts
-
-export const { GET, POST } = handlers; // Export the handlers
+export { GET, POST } from "@/auth";
